App refactoring to promote dry, reusable and consistent code
